egn_notifications/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
twilio==6.29.4
